backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
anthropic>=0.40.0
sqlalchemy>=2.0.0
python-multipart>=0.0.12
pillow>=11.0.0
python-dotenv>=1.0.0
pydantic>=2.9.0
tenacity>=9.0.0
aiofiles>=24.1.0
